Fix pr comments
